Riss: incremental improvements
This commits implements many improvements and fixes at once. The new ipasir
functionality to share new learned clauses is now supported, with a dynamic
allocation based on the parameter that is specified when initializing the
callback.

The refineFinalConflict function has been fixed, and now always jumps to
decision level 0 after a refinement attempt.

Incremental solving has been improved to (1) limit restarts to jump only
on top of the current assumptions, (2) allow continuing search with matching
assumptions between two consecutive calls, and (3) to prefetch assumptions by
checking whether a "selected" assumption of the full assumption set is already
falsified to abort a search early.

Modification (2) also required a new method to integrate clauses into the
solver in an intermediate search state with a given trail. This method adds
clauses lazily, meaning whenever unit propagation would be triggered by a new
clause, we decide to jump back one more level to not trigger propagation, as
this effort might be undone as soon as the next clause is added. Furthermore,
new assumptions have to be matched to old assumptions. This is done based on
a prefix match. Only the common prefix of new assumptions and assigned
assumptions is considered as the state to keep in an incremental call.