ipasir: configure Riss via environment
The solver configuration for Riss can now be specified via the environment
variable RISSCONFIG. If this variable is set, it's value is used to
initialize Riss when being used via the IPASIR interface.

As incremental calls might introduce new variables, and might reference
already given variables, this configuration has to be constructed with
care, because some simplification techniques do not support incremental
solving.