travis: initial commit
Continuous integration in the repository is run via the scripts/ci.sh
script. This script requires a recent version of astyle to check for
coding style. Furthermore, the fuzz checks require picosat to be
available on the system. Finally, a recent version of gcc should be
used.

This initial commit makes all these requirements available for a
single build configuration.