analyzeFinal: fix top level conflict
When an assertion on decision level 0 leads to a conflict, the previous
version of the code would have created an empty conflict, which would
turn the input formula into an unsatisfiable formula, independently of
the assumption literal. This fix makes sure that the responsible
assumption literal is always present in the conflict.