SMAC: first example run with Riss via script
The files that are added with this commit allow to easily run Riss together
with the algorithm configuration tool SMAC. Until now, the given files can
only be seen as an example on how to use SMAC to tune the parameters of Riss.

When running the file "./tools/smac-files/smac.sh", SMAC3 is installed with
its dependencies, and an example run of Riss is performed.