make: allow to build without git
Until now, git was required to generate the build information. However,
when downloading the plain files of the solver without the .git information
the generation process failed, because it was looking for git.

With this patch, the generated version information takes this case into
account, and writes "unavailable" instead of commit ID and date.